get product by id/updated model/routes
