+++ src/routes/Product.route.js
@@ -1,11 +1,12 @@
 import Router from 'express';
-import postProduct from '../controllers/Product.controller.js'
+import {postProduct, getProductById} from '../controllers/Product.controller.js'
 
 
 
 const productRoutes = Router();
 
 productRoutes.post('/post', postProduct)
+productRoutes.get('/detail/:id', getProductById)
 
 
 export default productRoutes;
